backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
sqlalchemy>=2.0.0
geoalchemy2>=0.13.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.6
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.28.2
celery>=5.3.0
openai>=1.1.1
anthropic>=0.5.0
spacy>=3.5.2
beautifulsoup4>=4.12.0
playwright>=1.35.0
scrapy>=2.8.0
httpx>=0.24.0
pytest>=7.3.1
